SDIT-3142 Added GET /visits/configuration/time-slots/prison-id/{prisonId}/day-of-week/{dayOfWeek}/time-slot-sequence/{timeSlotSequence} (#1447)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/jpa/repository/AgencyVisitTimeRepository.kt
@@ -44,5 +44,5 @@ interface AgencyVisitTimeRepository : JpaRepository<AgencyVisitTime, AgencyVisit
 interface VisitTimeSlotIdProjection {
   val prisonId: String
   val weekdayCode: String
-  val timeSlotSequence: Long
+  val timeSlotSequence: Int
 }
